fixed more warnings in rtgui (reported by gcc but not clang)

+++ rtgui/dirbrowser.cc
@@ -348,7 +348,7 @@ void DirBrowser::updateDir (const Gtk::TreeModel::iterator& iter)
     auto dir = Gio::File::create_for_path (iter->get_value (dtColumns.dirname));
     auto subDirs = listSubDirs (dir, options.fbShowHidden);
 
-    for (int i = 0; i < subDirs.size(); i++) {
+    for (size_t i = 0; i < subDirs.size(); i++) {
         bool found = false;
 
         for (Gtk::TreeModel::iterator it = iter->children().begin(); it != iter->children().end() && !found ; ++it) {
